# contracts/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/workflows/booking-corrections": {
      "post": {
        "operationId": "WorkflowsController_createCorrection",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create booking correction workflow request",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/shift-swaps": {
      "post": {
        "operationId": "WorkflowsController_createShiftSwap",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create shift swap workflow request",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/overtime-approvals": {
      "post": {
        "operationId": "WorkflowsController_createOvertimeApproval",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create overtime approval workflow request",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/inbox": {
      "get": {
        "operationId": "WorkflowsController_inbox",
        "parameters": [
          {
            "name": "status",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "overdueOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workflow inbox for authenticated approver/requester",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/policies": {
      "get": {
        "operationId": "WorkflowsController_policies",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workflow policies (HR/Admin)",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/policies/{type}": {
      "put": {
        "operationId": "WorkflowsController_upsertPolicy",
        "parameters": [
          {
            "name": "type",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create or update workflow policy (HR/Admin)",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/delegations": {
      "get": {
        "operationId": "WorkflowsController_delegations",
        "parameters": [
          {
            "name": "delegatorId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "workflowType",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List workflow delegation rules (HR/Admin)",
        "tags": [
          "workflows"
        ]
      },
      "post": {
        "operationId": "WorkflowsController_createDelegation",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create workflow delegation rule (HR/Admin)",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/delegations/{id}": {
      "patch": {
        "operationId": "WorkflowsController_updateDelegation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update workflow delegation rule (HR/Admin)",
        "tags": [
          "workflows"
        ]
      },
      "delete": {
        "operationId": "WorkflowsController_deleteDelegation",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete workflow delegation rule (HR/Admin)",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/{id}": {
      "get": {
        "operationId": "WorkflowsController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get workflow detail",
        "tags": [
          "workflows"
        ]
      }
    },
    "/v1/workflows/{id}/decision": {
      "post": {
        "operationId": "WorkflowsController_decide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Apply workflow action (approve/reject/delegate/cancel)",
        "tags": [
          "workflows"
        ]
      }
    },
...
    "/v1/terminal/sync/batches": {
      "post": {
        "operationId": "TerminalSyncController_importBatch",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import terminal offline-sync batch (CSV adapter v0)",
        "tags": [
          "terminal-sync"
        ]
      }
    },
    "/v1/terminal/sync/batches/file": {
      "post": {
        "operationId": "TerminalSyncController_importBatchFile",
        "parameters": [],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Import terminal offline-sync file batch (HONEYWELL_CSV_V1)",
        "tags": [
          "terminal-sync"
        ]
      }
    },
    "/v1/terminal/sync/batches/{id}": {
      "get": {
        "operationId": "TerminalSyncController_getBatch",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get imported terminal batch by id",
        "tags": [
          "terminal-sync"
        ]
      }
    },
...
    "/v1/policies": {
      "get": {
        "operationId": "PoliciesController_bundle",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Resolve active policy bundle for a given date",
        "tags": [
          "policy"
        ]
      }
    },
...
  },
  "info": {
    "title": "cueq API",
    "description": "Integrated time-tracking, absence-management, and shift-planning API for a German university (NRW / TV-L)",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
...
  }
}